Movidas funciones de cambios de DescargaTemporada a SMACB. Añadida muestra de cambios de calendario
